adex/requirements/base.txt
Django==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
django-cors-headers==3.2.0
whitenoise==5.0.1
